src/requirements/base.txt
asgiref==3.5.0
astroid==2.11.3
black==22.3.0
boto3==1.21.46
botocore==1.24.46
certifi==2021.10.8
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
drf-spectacular==0.22.0
gevent==21.8.0
gunicorn==20.1.0
isort==5.10.1
jmespath==1.0.0
# lazy-object-proxy==1.7.1
mccabe==0.7.0
psycopg2==2.9.3
pylint==2.13.7
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.0
